Factor out insubordination_scenarios, add small tests for components
